+++ app/src/main/java/co/touchlab/kampkit/android/MainActivity.kt
@@ -6,7 +6,7 @@ import androidx.activity.compose.setContent
 import co.touchlab.kampkit.android.ui.MainScreen
 import co.touchlab.kampkit.android.ui.theme.KaMPKitTheme
 import co.touchlab.kampkit.injectLogger
-import co.touchlab.kampkit.models.BreedViewModel
+import co.touchlab.kampkit.feature.breed.BreedViewModel
 import co.touchlab.kermit.Logger
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.koin.core.component.KoinComponent
