+++ shared/src/commonMain/kotlin/co/touchlab/kampkit/Koin.kt
@@ -5,8 +5,8 @@ import co.touchlab.kampkit.base.StaleDataDelegateImpl
 import co.touchlab.kampkit.ktor.Api
 import co.touchlab.kampkit.ktor.ApiImpl
 import co.touchlab.kampkit.ktor.HttpClientProvider
-import co.touchlab.kampkit.models.BreedRepository
-import co.touchlab.kampkit.models.BreedRepositoryImpl
+import co.touchlab.kampkit.feature.breed.BreedRepository
+import co.touchlab.kampkit.feature.breed.BreedRepositoryImpl
 import co.touchlab.kermit.Logger
 import co.touchlab.kermit.StaticConfig
 import co.touchlab.kermit.platformLogWriter
