+++ shared/src/commonMain/kotlin/co/touchlab/kampkit/ktor/Api.kt
@@ -1,10 +1,10 @@
 package co.touchlab.kampkit.ktor
 
 import co.touchlab.kampkit.base.ApiStatus
-import co.touchlab.kampkit.response.BreedResult
+import co.touchlab.kampkit.ktor.dto.BreedDto
 
 interface Api {
 
-    suspend fun getBreeds(): ApiStatus<BreedResult>
+    suspend fun getBreeds(): ApiStatus<BreedDto>
 
 }
