--- shared/src/commonMain/kotlin/co/touchlab/kampkit/models/BreedData.kt
@@ -1,7 +1,0 @@
-package co.touchlab.kampkit.models
-
-import co.touchlab.kampkit.db.Breed
-
-data class BreedData(
-    val breeds: List<Breed>
-)
