+++ shared/src/commonTest/kotlin/co/touchlab/kampkit/BreedRepositoryImplTest.kt
@@ -3,7 +3,7 @@ package co.touchlab.kampkit
 import co.touchlab.kampkit.base.StaleDataDelegate
 import co.touchlab.kampkit.db.Breed
 import co.touchlab.kampkit.ktor.Api
-import co.touchlab.kampkit.models.BreedRepositoryImpl
+import co.touchlab.kampkit.feature.breed.BreedRepositoryImpl
 import co.touchlab.kermit.Logger
 import co.touchlab.kermit.StaticConfig
 import kotlinx.coroutines.Dispatchers
