+++ shared/src/iosMain/kotlin/co/touchlab/kampkit/BreedCallbackViewModel.kt
@@ -1,8 +1,8 @@
 package co.touchlab.kampkit
 
 import co.touchlab.kampkit.db.Breed
-import co.touchlab.kampkit.models.BreedRepositoryImpl
-import co.touchlab.kampkit.models.BreedViewModel
+import co.touchlab.kampkit.feature.breed.BreedRepositoryImpl
+import co.touchlab.kampkit.feature.breed.BreedViewModel
 import co.touchlab.kampkit.models.CallbackViewModel
 import co.touchlab.kermit.Logger
